pyproject.toml
[project]
name = "zuno-bc-wallet"
version = "0.1.0"
description = "Zuno Embedded Wallet Service using Circle Developer-Controlled Wallets SDK"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "circle-developer-controlled-wallets>=8.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
]
